requirements.txt
python-telegram-bot
sqlalchemy
psycopg2-binary
python-dotenv

